modules/camera/requirements.txt
azure-iot-device~=2.7.0
requests
numpy
opencv-python-headless